Renamed the files for the nutrition label to be consistent with the project naming convention
