+++ src/components/nutrition/label/separator-bar.tsx
